fix on containers && card structure (#29)
