+++ package.json
@@ -3,22 +3,19 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.3.2",
-    "@testing-library/user-event": "^7.1.2",
-    "connected-react-router": "^6.8.0",
-    "history": "^4.10.1",
-    "node-sass": "^4.14.0",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
-    "react-hook-form": "^5.7.2",
-    "react-redux": "^7.2.0",
-    "react-router-dom": "^5.1.2",
-    "react-scripts": "3.4.1",
+    "react-router-dom": "^5.2.0",
+    "react-scripts": "3.4.3",
     "redux": "^4.0.5",
+    "react-redux": "^7.2.0",
+    "connected-react-router": "^6.8.0",
+    "history": "^4.10.1",
+    "node-sass": "^4.14.1",
     "redux-devtools-extension": "^2.13.8",
     "redux-ducks-ts": "^1.0.9",
     "redux-logger": "^3.0.6",
+    "react-hook-form": "^5.7.2",
     "redux-saga": "^1.1.3",
     "reselect": "^4.0.0"
   },
@@ -42,5 +39,8 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "typescript": "^4.0.2"
   }
 }
