+++ src/App.jsx
@@ -1,15 +1,12 @@
 import React from "react";
-import { Switch, Route, Redirect, BrowserRouter } from "react-router-dom";
+import { BrowserRouter } from "react-router-dom";
 // Redux:
 import { Provider } from "react-redux";
 import makeStore, { history } from "state";
 import { ConnectedRouter } from "connected-react-router";
 // Components
-import CreateCardView from "containers/CreateCard";
-import CardDetails from "containers/CardDetails";
-import HomeView from "containers/Home";
-import Search from "containers/Search";
-import Profile from "containers/Profile";
+import UserContainer from "./containers/User";
+
 /* import ThemeProvider from "components/_utils/ThemeProvider"; */
 
 const store = makeStore();
@@ -18,11 +15,7 @@ const App = () => (
 	<Provider store={store}>
 		<ConnectedRouter history={history}>
 			<BrowserRouter>
-				<Search />
-				<Profile />
-				<HomeView />
-				<CardDetails />
-				<CreateCardView />
+				<UserContainer />
 			</BrowserRouter>
 		</ConnectedRouter>
 	</Provider>
