--- src/components/Admin/MyComponent/index.scss
@@ -1,37 +1,0 @@
-.MyComponent {
-  text-align: center;
-  &-logo {
-    height: 40vmin;
-    pointer-events: none;
-  }
-  &-header {
-    background-color: #282c34;
-    min-height: 100vh;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    font-size: calc(10px + 2vmin);
-    color: white;
-  }
-  &-buttons{
-    margin: 1rem;
-  }
-  &-link {
-    color: #61dafb;
-  }
-  @media (prefers-reduced-motion: no-preference) {
-    &-logo {
-      animation: MyComponent-logo-spin infinite 20s linear;
-    }
-  }
-
-  @keyframes MyComponent-logo-spin {
-    from {
-      transform: rotate(0deg);
-    }
-    to {
-      transform: rotate(360deg);
-    }
-  }
-}
