+++ src/components/_shared/Card/CardContent/index.scss
