+++ src/components/_shared/Card/CardImage/index.scss
