+++ src/components/_shared/Card/ModalConfirmDelete/index.scss
