--- src/containers/CreateCard/index.js
@@ -1,18 +1,0 @@
-import React, { Suspense } from "react";
-import { Switch } from "react-router-dom";
-import routes from "./routes";
-
-// Components
-import MultipleRoutes from "components/_shared/MultipleRoutes";
-
-const CreateCard = () => (
-	<Suspense fallback={<span>loading....</span>}>
-		<Switch>
-			{routes.map((route, i) => (
-				<MultipleRoutes key={`${i}_${route.path}`} {...route} />
-			))}
-		</Switch>
-	</Suspense>
-);
-
-export default CreateCard;
