--- src/containers/Home/routes.js
@@ -1,12 +1,0 @@
-import { lazy } from "react";
-
-// Containers
-const Home = lazy(() => import("views/Home"));
-
-export default [
-	{
-		path: "/",
-		component: Home,
-		exact: true,
-	},
-];
