--- src/containers/Onboarding/routes.js
@@ -1,12 +1,0 @@
-import { lazy } from "react";
-
-// Containers
-const Onboarding = lazy(() => import("views/Onboarding"));
-
-export default [
-	{
-		path: "/onboarding",
-		component: Onboarding,
-		exact: true,
-	},
-];
